requirements.txt
torch
torchvision
matplotlib
numpy
opencv-python
imageio
imageio-ffmpeg
pyyaml
tqdm
argparse
cython
scikit-image
scipy
onnxruntime

future-fstrings
tensorflow-gpu